Fix Color565 for both 8 and 16-bit mode
Red and Blue were reversed in Color565.
8-bit mode requires a different calculation for blue. It's not obvious
from the datasheet if this is a bug.
Wasn't identified as a keyword.
Also adjusted colour presets since "pink" looked like magenta in 8-bit
mode and like red in 16-bit mode. Purple is now more distinct from blue
and consistent in both modes.